Moved deep_get to a util file
