+++ util_ships.py
@@ -1,6 +1,7 @@
 """Utility functions relating to ships."""
 from edmc_data import ship_name_map
 
+
 def ship_file_name(ship_name: str, ship_type: str) -> str:
     """Return a ship name suitable for a filename."""
     name = str(ship_name or ship_name_map.get(ship_type.lower(), ship_type)).strip()
